ambot/requirements.txt
cryptography>=41.0.0
python-dotenv>=1.0.0
click>=8.1.0 